extract persistence code to another module
